**Faster unit tests.**

- [X] Include this change in the Release Notes. If checked, you must provide an **end-user friendly** description for your change below

---
We should test only our own code in unit tests. (If we need end to end testing, integration tests is the place for this). By effective use of mocks to test only our own code, we speed up our tests and therefore our builds.  Local changes have shown this takes ~15 seconds of the test time (When using a single worker).

I'd like to expand this out to our other tests. After this, the slowest are (measured using --duration):
```
5.25s setup    tests/unit/test_memory.py::test_add_event_with_action
3.28s call     tests/unit/test_security.py::test_unsafe_python_code
3.11s call     tests/unit/test_memory.py::test_add_event_increment_thought_idx
2.41s call     tests/unit/test_security.py::test_unsafe_bash_command
2.02s call     tests/unit/test_llm.py::test_completion_retries[InternalServerError-extra_args1-2]
2.02s call     tests/unit/test_llm.py::test_completion_retries[ServiceUnavailableError-extra_args2-2]
2.02s call     tests/unit/test_llm.py::test_completion_retries[RateLimitError-extra_args3-2]
2.01s call     tests/unit/test_llm.py::test_completion_retries[APIConnectionError-extra_args0-2]
2.01s call     tests/unit/test_llm.py::test_completion_exhausts_retries
2.00s call     tests/unit/test_runtime_build.py::test_build_image_from_scratch
1.46s setup    tests/unit/test_memory.py::test_embeddings_inserted_into_chroma
1.42s setup    tests/unit/test_memory.py::test_add_event_with_missing_keys
1.41s call     tests/unit/test_agent_skill.py::test_find_file_not_exist_file
1.35s setup    tests/unit/test_memory.py::test_load_events_batch_insert
1.35s setup    tests/unit/test_memory.py::test_add_event_with_observation
1.34s setup    tests/unit/test_memory.py::test_add_event_increment_thought_idx
1.25s setup    tests/unit/test_memory.py::test_search_with_no_results
1.16s setup    tests/unit/test_memory.py::test_search_returns_correct_results
1.14s setup    tests/unit/test_memory.py::test_load_events_into_index_with_invalid_json
1.13s setup    tests/unit/test_memory.py::test_events_to_docs_no_events
1.04s call     tests/unit/test_acompletion.py::test_async_completion_with_user_cancellation[0.9]
```