@tobitege 2024-10-15
One way to make unit tests much fast without changing them is to run pytest with e.g. "-n 4" to run 4 tests in parallel within the workflow.
Like: `poetry run pytest -v -n 4 --reruns 1 --reruns-delay 3 ./tests/test_*.py`