@tofarr 2024-10-15
> One way to make unit test workflow finish much faster without changing them is to run pytest with e.g. "-n 4" to run 4 tests in parallel within the workflow. Like: `poetry run pytest -v -n 4 --reruns 1 --reruns-delay 3 ./tests/test_*.py`

For sure - multiple workers make things faster.